+++ dbbackup/management/commands/dbbackup.py
@@ -101,6 +101,11 @@ class Command(BaseDbBackupCommand):
             "connector": f"{self.connector.__module__}.{self.connector.__class__.__name__}",
         }
         metadata_filename = f"{filename}.metadata"
+
+        # Load custom metadata if configured
+        user_metadata = utils.get_user_metadata(metadata)
+        metadata.update(user_metadata)
+
         metadata_content = json.dumps(metadata)
 
         if local: